+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+### Fixed
+- Respect browser and system proxy settings when noturno is off
 
 ## [2.0.0] 05-01-2020
 ### Added