+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased]
+## [2.1.1] 05-10-2020
+### Fixed
+- Use safe loads when injecting HTML content
 
 ## [2.1.0] 05-10-2020
 ### Fixed